refactor: organização do código
