+++ backend/medqueue/src/main/java/com/medqueue/medqueue/controller/paciente/InfoFilaPacienteController.java
@@ -1,17 +1,11 @@
 package com.medqueue.medqueue.controller.paciente;
 
-
-import java.util.List;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.medqueue.medqueue.dto.FilaPacienteDTO;
 import com.medqueue.medqueue.dto.InfoFilaParaPacienteDTO;
-import com.medqueue.medqueue.models.FilaPaciente;
 import com.medqueue.medqueue.service.paciente.InfoFilaPacienteService;
 
 import io.swagger.v3.oas.annotations.Operation;
