+++ backend/medqueue/src/main/java/com/medqueue/medqueue/security/JwtRequestFilter.java
@@ -41,6 +41,11 @@ public class JwtRequestFilter extends OncePerRequestFilter {
         String username = null;
         String jwt = null;
 
+        if (path.startsWith("/api/auth")) {
+            chain.doFilter(request, response);
+            return;
+        }
+
         if (authHeader != null && authHeader.startsWith("Bearer ")) {
             jwt = authHeader.substring(7);
 
@@ -62,11 +67,6 @@ public class JwtRequestFilter extends OncePerRequestFilter {
                 authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
                 SecurityContextHolder.getContext().setAuthentication(authToken);
             }
-        }
-
-        if (path.startsWith("/api/auth")) {
-            chain.doFilter(request, response);
-            return;
         }
 
         chain.doFilter(request, response);
